:recycle: Move spacing file and apply switch statement
